+++ conans/client/manager.py
@@ -40,7 +40,7 @@ class ConanManager(object):
     """ Manage all the commands logic  The main entry point for all the client
     business logic
     """
-    def __init__(self, client_cache, user_io, runner, remote_manager, search_manager):
+    def __init__(self, client_cache, user_io, runner, remote_manager, search_manager, tools):
         assert isinstance(user_io, UserIO)
         assert isinstance(client_cache, ClientCache)
         self._client_cache = client_cache
@@ -49,6 +49,7 @@ class ConanManager(object):
         self._remote_manager = remote_manager
         self._current_scopes = None
         self._search_manager = search_manager
+        self._tools = tools
 
     def export(self, user, conan_file_path, keep_source=False, filename=None):
         """ Export the conans
@@ -92,7 +93,7 @@ class ConanManager(object):
         remote_proxy = ConanProxy(self._client_cache, self._user_io, self._remote_manager,
                                   remote_name=None, update=False, check_updates=False,
                                   manifest_manager=None)
-        loader = ConanFileLoader(self._runner, self._client_cache.settings, profile)
+        loader = ConanFileLoader(self._runner, self._tools, self._client_cache.settings, profile)
         conanfile = loader.load_virtual([reference], current_path)
         graph_builder = self._get_graph_builder(loader, False, remote_proxy)
         deps_graph = graph_builder.load(conanfile)
@@ -171,7 +172,7 @@ class ConanManager(object):
         remote_proxy = ConanProxy(self._client_cache, self._user_io, self._remote_manager, remote,
                                   update=False, check_updates=check_updates)
 
-        loader = ConanFileLoader(self._runner, self._client_cache.settings, profile)
+        loader = ConanFileLoader(self._runner, self._tools, self._client_cache.settings, profile)
         conanfile = self._get_conanfile_object(loader, reference, filename, current_path)
         graph_builder = self._get_graph_builder(loader, False, remote_proxy)
         deps_graph = graph_builder.load(conanfile)
@@ -254,7 +255,7 @@ class ConanManager(object):
                                            interactive=manifest_interactive) if manifest_folder else None
         remote_proxy = ConanProxy(self._client_cache, self._user_io, self._remote_manager, remote,
                                   update=update, check_updates=False, manifest_manager=manifest_manager)
-        loader = ConanFileLoader(self._runner, self._client_cache.settings, profile)
+        loader = ConanFileLoader(self._runner, self._tools, self._client_cache.settings, profile)
         conanfile = self._get_conanfile_object(loader, reference, filename, current_path)
         graph_builder = self._get_graph_builder(loader, update, remote_proxy)
         deps_graph = graph_builder.load(conanfile)
@@ -312,7 +313,7 @@ class ConanManager(object):
             conanfile_path = os.path.join(reference, CONANFILE)
             conanfile = load_consumer_conanfile(conanfile_path, current_path,
                                                 self._client_cache.settings, self._runner,
-                                                output)
+                                                output, self._tools)
             export_folder = reference
             config_source_local(export_folder, current_path, conanfile, output)
         else:
@@ -320,7 +321,7 @@ class ConanManager(object):
             conanfile_path = self._client_cache.conanfile(reference)
             conanfile = load_consumer_conanfile(conanfile_path, current_path,
                                                 self._client_cache.settings, self._runner,
-                                                output, reference)
+                                                output, self._tools, reference)
             src_folder = self._client_cache.source(reference, conanfile.short_paths)
             export_folder = self._client_cache.export(reference)
             config_source(export_folder, src_folder, conanfile, output, force)
@@ -342,7 +343,8 @@ class ConanManager(object):
 
         conanfile = load_consumer_conanfile(conan_file_path, current_path,
                                             self._client_cache.settings,
-                                            self._runner, output, reference, error=True)
+                                            self._runner, output, self._tools,
+                                            reference, error=True)
 
         if dest_folder:
             if not os.path.isabs(dest_folder):
@@ -360,7 +362,7 @@ class ConanManager(object):
         conan_file_path = os.path.join(build_folder, CONANFILE)
         conanfile = load_consumer_conanfile(conan_file_path, current_path,
                                             self._client_cache.settings,
-                                            self._runner, output)
+                                            self._runner, output, self._tools)
         packager.create_package(conanfile, build_folder, current_path, output, local=True)
 
     def package(self, reference, package_id):
@@ -400,7 +402,7 @@ class ConanManager(object):
             output.info("Re-packaging %s" % package_reference.package_id)
             conanfile = load_consumer_conanfile(conan_file_path, build_folder,
                                                 self._client_cache.settings,
-                                                self._runner, output, reference)
+                                                self._runner, output, self._tools, reference)
             rmdir(package_folder)
             with environment_append(conanfile.env):
                 packager.create_package(conanfile, build_folder, package_folder, output)
@@ -423,7 +425,7 @@ class ConanManager(object):
             output = ScopedOutput("Project", self._user_io.out)
             conan_file = load_consumer_conanfile(conan_file_path, current_path,
                                                  self._client_cache.settings,
-                                                 self._runner, output)
+                                                 self._runner, output, self._tools)
         except NotFoundException:
             # TODO: Auto generate conanfile from requirements file
             raise ConanException("'%s' file is needed for build.\n"
