+++ conans/test/util/conanfile_tools_test.py
@@ -1,16 +1,16 @@
-import unittest
 import os
-from conans.util.files import save, load
-from conans.client.loader import ConanFileLoader
-from conans.model.settings import Settings
-from conans.model.options import OptionsValues
-from conans.test.utils.test_files import temp_folder
-from conans.model.scope import Scopes
-from conans import tools
-from nose_parameterized.parameterized import parameterized
-from conans.test.utils.tools import TestClient
-from conans.model.profile import Profile
+import unittest
 
+from nose_parameterized.parameterized import parameterized
+
+from conans import tools
+from conans.client.loader import ConanFileLoader
+from conans.model.profile import Profile
+from conans.model.settings import Settings
+from conans.test.utils.test_files import temp_folder
+from conans.test.utils.tools import TestClient
+from conans.tools_factory import ToolsFactory
+from conans.util.files import save, load
 
 base_conanfile = '''
 from conans import ConanFile
@@ -117,7 +117,7 @@ class ConanfileToolsTest(unittest.TestCase):
         return tmp_dir, file_path, text_file
 
     def _build_and_check(self, tmp_dir, file_path, text_file, msg):
-        loader = ConanFileLoader(None, Settings(), Profile())
+        loader = ConanFileLoader(None, ToolsFactory.new(), Settings(), Profile())
         ret = loader.load_conan(file_path, None)
         curdir = os.path.abspath(os.curdir)
         os.chdir(tmp_dir)
